fixing mistakes in all classes
